Fix bug where closing connection closes already closed PGSQLPreparedStatement.
Remove flawed `PGSQLPreparedStatement::close()` method. This method is nearly identical
to `PGSQLStatement::close()`, however, it lacks a call to `conn.onStatementClosed(this)`.
Due to this deficiency, the connection will attempt to close the statement a 2nd time
when it is closed, resulting in an exception being thrown.

By erasing the implementation in `PGSQLPreparedStatement`, the correct implemention
in `PGSQLStatement` can be used.